Add files via upload
This Python file contains the implementation of an Artificial Neural Network (ANN) model aimed at predicting the probability of a patient having heart disease or experiencing a heart attack. The model is trained on a balanced dataset of health indicators and achieves an accuracy of 77% with an AUC of 0.816.

Key features of this model include:

Balanced precision and recall for accurate classification.
Visualization of the Precision-Recall curve to assess model performance.